Extract PartnersPage component
