+++ pages/partners.tsx
@@ -1,15 +1,8 @@
 import { PortalPartner } from "lib/portal-types";
-import type { NextPage, GetStaticProps } from "next";
-import { useMemo, useState } from "react";
-import strings from "content/strings.json";
-import FinancialPartners from "components/partners/sections/financial";
-import ExpertsPartners from "components/partners/sections/experts";
-import BecomePartner from "components/partners/sections/become-partner";
-import { Layout, Section, SectionContent } from "components/layout";
-import * as S from "components/partners/styles";
-import Tabs from "components/partners/tabs";
+import { NextPage, GetStaticProps } from "next";
 import { siteData } from "lib/site-data";
 import { Article } from "lib/related-blog-posts";
+import PartnersPage from "components/partners/page";
 
 type PageProps = {
   partners: readonly PortalPartner[];
@@ -17,53 +10,7 @@ type PageProps = {
 };
 
 const Page: NextPage<PageProps> = ({ partners, blogPosts }) => {
-  const msg = strings.pages.partners;
-  const sections = [
-    {
-      key: "financial",
-      label: msg.tabs.financial.title,
-      component: (
-        <FinancialPartners partners={partners} blogPosts={blogPosts} />
-      ),
-    },
-    {
-      key: "experts",
-      label: msg.tabs.experts.title,
-      component: <ExpertsPartners partners={partners} />,
-    },
-  ];
-
-  const [activeSectionKey, setActiveSectionKey] = useState<string>(
-    sections[0].key
-  );
-
-  const ActiveSection = useMemo(() => {
-    const componentToRender = sections.find(
-      ({ key }) => key === activeSectionKey
-    );
-
-    return componentToRender?.component ?? FinancialPartners;
-  }, [activeSectionKey]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  return (
-    <Layout
-      crumbs={[{ label: msg.navigation.partners }]}
-      head={{
-        title: msg.metadata.title,
-        description: msg.metadata.description,
-      }}
-    >
-      <Section>
-        <SectionContent>
-          <S.Heading>{msg.title}</S.Heading>
-          <S.Tagline>{msg.description}</S.Tagline>
-          <Tabs items={sections} onChange={setActiveSectionKey} />
-        </SectionContent>
-      </Section>
-      {ActiveSection}
-      <BecomePartner />
-    </Layout>
-  );
+  return <PartnersPage partners={partners} blogPosts={blogPosts} />;
 };
 
 export const getStaticProps: GetStaticProps<PageProps> = async () => {
